Added RootGPCA initialization
RootGPCA initialization substracts the minimum per time series to make
them positive. This is equivalent to modifying the last column of A. It
then takes the roots and hopes for them to lie on “d” subspaces  that
GPCA hopefully recovers. This only works if one of the eigenvalues of
modified A_i’s is dominant.